@@ -8,24 +8,32 @@
     "start": "next start",
     "lint": "next lint",
     "storybook": "start-storybook -p 6006 -s ./public",
-    "build-storybook": "build-storybook -s public"
+    "build-storybook": "build-storybook -s public",
+    "build-dictionary": "node ./style-dictionary/build.js"
   },
   "dependencies": {
     "@headlessui/react": "^1.7.15",
     "@heroicons/react": "^2.0.18",
+    "@reduxjs/toolkit": "^2.0.1",
     "@storybook/addon-postcss": "^2.0.0",
+    "axios": "^1.6.2",
     "classnames": "^2.3.1",
     "iconsax-react": "^0.0.8",
-    "next": "12.3.0",
+    "jsonwebtoken": "^9.0.2",
+    "next": "^14.0.4",
     "postcss-loader": "^7.0.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
+    "react-redux": "^9.0.2",
     "react-slick": "^0.28.1",
+    "react-spinners": "^0.13.8",
+    "react-toastify": "^9.1.3",
     "rsuite": "^5.16.4",
     "sb": "^6.5.13",
     "slick-carousel": "^1.8.1",
-    "storybook-addon-next": "^1.6.10"
+    "storybook-addon-next": "^1.6.10",
+    "tailwind-merge": "^2.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.2",
@@ -38,9 +46,12 @@
     "@storybook/manager-webpack5": "^6.5.13",
     "@storybook/react": "^6.5.13",
     "@storybook/testing-library": "^0.0.13",
+    "@types/jsonwebtoken": "^9.0.5",
     "@types/node": "18.7.18",
     "@types/react": "18.0.20",
     "@types/react-dom": "18.0.6",
+    "@types/react-redux": "^7.1.32",
+    "@types/react-slick": "^0.23.13",
     "autoprefixer": "^10.4.11",
     "babel-loader": "^8.3.0",
     "classnames": "^2.3.1",
@@ -50,6 +61,7 @@
     "postcss": "^8.4.16",
     "postcss-preset-env": "^7.8.2",
     "react-hook-form": "^7.26.0",
+    "style-dictionary": "^3.8.0",
     "tailwindcss": "^3.1.8",
     "typescript": "4.8.3",
     "yup": "^0.32.11"
