@@ -1,22 +1,17 @@
 import React from "react";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { useRouter, usePathname } from "next/navigation";
 import classNames from "classnames";
 
 type NavLinkProp = {
-  containerClassName: string;
-  icon: JSX.Element | React.ReactNode;
+  containerClassName?: string;
+  Icon: any;
   name: string;
   url: string;
 };
-const NavLink = ({
-  containerClassName,
-  icon,
-  name,
-  url,
-}: NavLinkProp) => {
-  const { asPath } = useRouter();
-  const isActiveIcon = url === asPath;
+const NavLink = ({ containerClassName, Icon, name, url }: NavLinkProp) => {
+  const pathName = usePathname();
+  const isActiveIcon = url === pathName;
   return (
     <li
       className={classNames(
@@ -27,25 +22,25 @@ const NavLink = ({
         containerClassName
       )}
     >
-      <Link href={url} passHref>
+      <Link href={url} passHref legacyBehavior>
         <a className="inline-flex items-center w-full text-sm font-normal font-Inter text-secondary-700 focus:text-secondary-700 transition-colors duration-150 hover:text-primary-500 ">
           <span
             className={classNames(
-              { "text-secondary-800": icon && !isActiveIcon },
-              { "text-primary-500": icon && isActiveIcon }
+              { "text-secondary-800": Icon && !isActiveIcon },
+              { "text-primary-500": Icon && isActiveIcon }
             )}
           >
-            {icon}
+            <Icon size="22" variant="Bulk" />
           </span>
-         <span className="ml-4">{name}</span>
+          <span className="ml-4">{name}</span>
         </a>
       </Link>
     </li>
   );
 };
 
-NavLink.defaultProps = {
-  containerClassName: null,
-};
+// NavLink.defaultProps = {
+//   containerClassName: null,
+// };
 
 export default NavLink;
