@@ -21,7 +21,7 @@ const AuthPage = ({ children, image, imagealt, text, title }: AuthPages) => {
             <div className="flex flex-col justify-center flex-1 px-4 py-12 sm:px-6 lg:flex-none lg:px-20 xl:px-24 ">
               <div className="w-full max-w-[43rem] lg:w-[38rem]  mx-auto pt-5 lg:pt-0 ">
                 <div className="relative lg:py-10 lg:px-8 lg:my-4 lg:max-w-[38rem] lg:w-[580px]">
-                  <Link passHref href="/">
+                  <Link passHref href="/" legacyBehavior>
                     <a className=" hidden lg:block absolute left-3 top-[-8px] cursor-pointer">
                       <Image
                         src="/assets/signUp/logo-black.svg"
@@ -33,7 +33,7 @@ const AuthPage = ({ children, image, imagealt, text, title }: AuthPages) => {
                   </Link>
                   <div className="bg-secondary-25  p-6 lg:p-8 ">
                     <div className="mb-8">
-                      <h2 className=" font-RozhaOne max-w-[300px] md:max-w-[424px] text-2xl md:text-4xl font-extrabold lg:font-bold text-tertiary pb-2">
+                      <h2 className=" font-RozhaOne max-w-[300px] md:max-w-[424px] text-2xl md:text-4xl font-extrabold lg:font-bold text-neutral-tertiary pb-2">
                         {title}
                       </h2>
                       <p className="text-secondary-700 text-[14px]">{text}</p>
@@ -44,8 +44,16 @@ const AuthPage = ({ children, image, imagealt, text, title }: AuthPages) => {
               </div>
             </div>
             {image && (
-              <div className="relative lg:flex flex-1 hidden lg:w-full bottom-0 ">
-                <Image src={image} alt={imagealt} />
+              <div className="relative lg:flex flex-1 hidden lg:w-full lg:min-h-screen bottom-0 ">
+                <Image
+                  src={image}
+                  alt={imagealt}
+                  fill
+                  sizes="100vw"
+                  // layout="fill"
+                  // objectFit="cover"
+                  // className="min-h-screen w-full"
+                />
               </div>
             )}
           </div>
@@ -55,10 +63,10 @@ const AuthPage = ({ children, image, imagealt, text, title }: AuthPages) => {
   );
 };
 
-AuthPage.defaultProps = {
-  children: null,
-  image: null,
-  imagealt: "authencation page image",
-};
+// AuthPage.defaultProps = {
+//   children: null,
+//   image: null,
+//   imagealt: "authencation page image",
+// };
 
 export default AuthPage;
