@@ -20,13 +20,13 @@ const BlogCard = ({ article, slug, src, pressRelease }: BlogCards) => {
           alt="blogs images"
         />
       </div>
-      <div className="absolute  inset-0 px-6 py-4 transition-opacity duration-200 bg-black bg-opacity-[0.6] opacity-0 hover:opacity-100">
+      <div className="absolute  inset-0 px-6 py-4 transition-opacity duration-200 bg-neutral-black bg-opacity-[0.6] opacity-0 hover:opacity-100">
         <div className="absolute bottom-5">
-          <p className="mb-2  text-base font-normal text-white">
+          <p className="mb-2  text-base font-normal text-neutral-white">
             PRESS RELEASE . {pressRelease}{" "}
           </p>
-          <p className="text-sm font-Inter tracking-wide font-bold text-white">
-            <Link href={`/article/${slug}`} passHref>
+          <p className="text-sm font-Inter tracking-wide font-bold text-neutral-white">
+            <Link href={`/article/${slug}`} passHref legacyBehavior>
               {article}
             </Link>
           </p>
