@@ -11,7 +11,8 @@ const NavLink = ({ title, url }: Navlinks) => {
       <Link
         passHref
         href={url}
-        className="font-medium tracking-wide !text-black transition-colors duration-200 font-Inter text-sm"
+        className="font-medium tracking-wide !text-neutral-black transition-colors duration-200 font-Inter text-sm"
+        legacyBehavior
       >
         {title}
       </Link>
