@@ -7,13 +7,13 @@ const AttentionCard = () => {
   return (
     <div className="p-2">
       <div className="relative bg-primary-500 h-[300px] rounded-md w-full  bg-contain md:bg-none lg:bg-[url('/assets/dashboard/Vector.png')] bg-bottom bg-no-repeat bg-[url('/assets/dashboard/Vector.png')] p-7">
-        <div className="flex items-center text-white justify-between cursor-pointer">
+        <div className="flex items-center text-neutral-white justify-between cursor-pointer">
           <p className="text-sm ">Today</p>
           <Receipt size="24" variant="Outline" />
         </div>
         <div className="py-3">
           <div className="flex items-center mt-4 -space-x-3">
-            <span className=" flex items-center justify-center border-2 rounded-full hover:z-20 border-white">
+            <span className=" flex items-center justify-center border-2 rounded-full hover:z-20 border-neutral-white">
               <Image
                 className=" rounded-full "
                 width={40}
@@ -23,7 +23,7 @@ const AttentionCard = () => {
                 alt="user avatar3"
               />
             </span>
-            <span className="flex items-center justify-center border-2 rounded-full z-10 hover:z-20 border-white">
+            <span className="flex items-center justify-center border-2 rounded-full z-10 hover:z-20 border-neutral-white">
               <Image
                 className="rounded-full"
                 width={40}
@@ -33,7 +33,7 @@ const AttentionCard = () => {
                 alt="user avatar4"
               />
             </span>
-            <span className="flex items-center justify-center border-2 rounded-full z-10 hover:z-20 border-white">
+            <span className="flex items-center justify-center border-2 rounded-full z-10 hover:z-20 border-neutral-white">
               <Image
                 className="m-auto rounded-full"
                 width={40}
@@ -43,7 +43,7 @@ const AttentionCard = () => {
                 alt="user avatar5"
               />
             </span>
-            <span className="flex items-center justify-center border-2 rounded-full z-10 hover:z-20 border-white">
+            <span className="flex items-center justify-center border-2 rounded-full z-10 hover:z-20 border-neutral-white">
               <Image
                 className="m-auto rounded-full"
                 width={40}
@@ -54,12 +54,12 @@ const AttentionCard = () => {
               />
             </span>
           </div>
-          <div className=" text-white mt-3 ">
+          <div className=" text-neutral-white mt-3 ">
             <p className=" text-xl font-semibold">Ajo with Friend</p>
             <p className="text-sm">Payment of your turn with Frank and Jamal</p>
           </div>
         </div>
-        <Link href="" passHref>
+        <Link href="" passHref legacyBehavior>
           <a>
             <p className="text-sm text-secondary-100 hover:text-secondary-200 border-b border-secondary-100 absolute bottom-8">
               View Details
