@@ -1,39 +1,28 @@
 import React from "react";
 import classNames from "classnames";
 import Link from "next/link";
-
+import { styles } from "./NewActionCard.styles";
+import { ColorType } from "@/utils/colors";
 type NewActionCardProp = {
-  backGround: string;
-  borderColor: string;
+  color: ColorType;
   icon: JSX.Element | React.ReactNode;
   title: string;
-  textColor: string;
   url: string;
 };
 const NewActionCard = ({
-  backGround,
-  borderColor,
+  color,
   icon,
-  textColor,
+
   title,
   url,
 }: NewActionCardProp) => {
   return (
-    <div
-      className={classNames(
-        " w-1/2 lg:w-1/4 border m-0.5 flex flex-col  items-center justify-center transition duration-500 hover:scale-105 cursor-pointer rounded-md  h-40 lg:h-52",
-        backGround,
-        borderColor
-      )}
-    >
-      <Link href={url} passHref>
-        <a className="flex flex-col items-center justify-center">
-          <span className={textColor}>{icon}</span>
+    <div className={classNames(styles.base, styles.variant.background[color])}>
+      <Link href={url} passHref legacyBehavior>
+        <a className={classNames(styles.linkContainer)}>
+          <span className={classNames(styles.variant.text[color])}>{icon}</span>
           <p
-            className={classNames(
-              "w-32 text-center text-sm font-bold mt-3",
-              textColor
-            )}
+            className={classNames(styles.textColor, styles.variant.text[color])}
           >
             {title}
           </p>
