@@ -1,11 +1,38 @@
 import Image from "next/image";
 import React from "react";
 import Button from "@/components/form/Button";
+import userService from "../../../services/user";
+import connectionService from "../../../services/connection";
+import { useAppDispatch } from "../../../redux/hooks";
+import { useAppSelector } from "../../../redux/hooks";
+import { toast, ToastContainer } from "react-toastify";
 
-const NewFriendToast = () => {
+const NewFriendToast = ({ name, connectionId, successCallback }: any) => {
+  const accept = async () => {
+    try {
+      await connectionService.accept(connectionId);
+      // console.log(data);
+      toast.success("Request accepted");
+      successCallback(connectionId);
+    } catch (error: any) {
+      toast.error(error.message);
+    }
+  };
+
+  const decline = async () => {
+    try {
+      await connectionService.decline(connectionId);
+      toast.success("Request rejected");
+      successCallback(connectionId);
+      // console.log(data);
+    } catch (error: any) {
+      toast.error(error.message);
+    }
+  };
+
   return (
     <div
-      className="p-3 pl-6 w-full border hidden lg:flex border-secondary-50  bg-white rounded-lg shadow "
+      className="p-3 pl-6 w-full border hidden lg:flex border-secondary-50  bg-neutral-white rounded-lg shadow "
       role="alert"
     >
       <div className="flex items-center justify-between w-full">
@@ -19,25 +46,32 @@ const NewFriendToast = () => {
               objectFit="cover"
               alt="user image"
             />
-            <span className="inline-flex absolute right-0 bottom-1 border border-white justify-center items-center w-3 h-3 bg-green-600 rounded-full">
+            <span className="inline-flex absolute right-0 bottom-1 border border-neutral-white justify-center items-center w-3 h-3 bg-green-600 rounded-full">
               <span className="sr-only">active user</span>
             </span>
           </div>
           <div className="ml-3">
             <p className="text-sm font-medium font-Inter text-secondary-800">
-              Jumoke Philips wants to join your network{" "}
+              {/* Jumoke Philips wants to join your network */}
+              {name} wants to join your network{" "}
             </p>
           </div>
         </div>
         <div>
-          <Button color="light" size="xs" className="">
+          <Button
+            onClick={decline}
+            color="neutral-light"
+            size="xs"
+            className=""
+          >
             Decline
           </Button>
-          <Button className="ml-1" size="xs">
+          <Button onClick={accept} className="ml-1" size="xs">
             Accept
           </Button>
         </div>
       </div>
+      <ToastContainer />
     </div>
   );
 };
