@@ -9,12 +9,12 @@ type WidgetCardProps = {
 const WidgetCard = ({ children, title, url }: WidgetCardProps) => {
   return (
     <div className="py-2 m-2 ">
-      <div className="bg-white border border-secondary-50 rounded-md ">
+      <div className="bg-neutral-white border border-secondary-50 rounded-md ">
         <div className="flex items-center border-b border-secondary-50 justify-between p-6">
           <p className="text-base font-semibold leading-6 text-secondary-800">
             {title}
           </p>
-          <Link href={url}>
+          <Link href={url} legacyBehavior>
             <a>
               <p className="text-xs leading-normal tracking-wide font-bold cursor-pointer text-right text-primary-500">
                 See all
