@@ -1,19 +1,22 @@
 import React from "react";
 import classNames from "classnames";
 import { ButtonProps, Sizes, Styles } from "../../types/index";
+import { ClipLoader } from "react-spinners";
 
 const Button = ({
-  className,
+  className = "",
   children,
-  color,
+  color = "primary",
   disabled,
   icon,
-  leftIcon,
-  onClick,
+  leftIcon = null,
+  onClick = () => {},
   outline,
-  rightIcon,
-  size,
-}: ButtonProps) => {
+  rightIcon = null,
+  size = "sm",
+  type,
+  loading,
+}: ButtonProps | any) => {
   const sizes: Sizes = {
     xs: "py-1.5 px-3 text-xs",
     sm: "py-2 px-[1.125rem] text-sm",
@@ -25,31 +28,31 @@ const Button = ({
   const styles: Styles = {
     solid: {
       primary:
-        "text-white bg-primary-500 hover:bg-primary-600 hover:text-white focus:shadow-primary-xs ",
+        "text-neutral-white bg-primary-500 hover:bg-primary-600 hover:text-neutral-white focus:shadow-primary-xs ",
       secondary:
-        "text-white bg-secondary-800 hover:bg-secondary-900 hover:text-white focus:shadow-secondary-xs",
+        "text-neutral-white bg-secondary-800 hover:bg-secondary-900 hover:text-neutral-white focus:shadow-secondary-xs",
       light:
         "text-secondary-500 bg-secondary-50 hover:bg-secondary-100 hover:text-secondary-600 focus:shadow-secondary-xs",
       danger:
-        "text-white bg-danger-600 hover:bg-danger-500 hover:text-white focus:shadow-danger-xs",
+        "text-neutral-white bg-danger-600 hover:bg-danger-500 hover:text-neutral-white focus:shadow-danger-xs",
       warning:
-        "text-white bg-warning-500 hover:bg-warning-600 hover:text-white focus:shadow-warning-xs",
+        "text-neutral-white bg-warning-500 hover:bg-warning-600 hover:text-neutral-white focus:shadow-warning-xs",
       success:
-        "text-white bg-success-600 hover:bg-success-500 hover:text-white focus:shadow-success-xs",
+        "text-neutral-white bg-success-600 hover:bg-success-500 hover:text-neutral-white focus:shadow-success-xs",
     },
     outline: {
       primary:
-        "text-secondary-600 border border-secondary-600 hover:bg-secondary-600 hover:border-secondary-600 hover:text-white focus:shadow-primary-xs",
+        "text-secondary-600 border border-secondary-600 hover:bg-secondary-600 hover:border-secondary-600 hover:text-neutral-white focus:shadow-primary-xs",
       secondary:
-        "text-primary-500 border border-primary-500 hover:bg-secondary-600 hover:border-secondary-600 hover:text-white focus:shadow-secondary-xs ",
+        "text-primary-500 border border-primary-500 hover:bg-secondary-600 hover:border-secondary-600 hover:text-neutral-white focus:shadow-secondary-xs ",
       light:
         "text-secondary-500 bg-secondary-50 hover:bg-secondary-100 hover:text-secondary-600 focus:shadow-secondary-xs",
       danger:
-        "text-danger-600 border border-danger-600 hover:bg-danger-500 hover:border-danger-500 hover:text-white focus:shadow-danger-xs ",
+        "text-danger-600 border border-danger-600 hover:bg-danger-500 hover:border-danger-500 hover:text-neutral-white focus:shadow-danger-xs ",
       warning:
-        "text-warning-600 border border-warning-600 hover:bg-warning-500 hover:border-warning-500 hover:text-white focus:shadow-warning-xs ",
+        "text-warning-600 border border-warning-600 hover:bg-warning-500 hover:border-warning-500 hover:text-neutral-white focus:shadow-warning-xs ",
       success:
-        "text-success-600 border border-success-600 hover:bg-success-500 hover:border-success-500 hover:text-white focus:shadow-success-xs ",
+        "text-success-600 border border-success-600 hover:bg-success-500 hover:border-success-500 hover:text-neutral-white focus:shadow-success-xs ",
     },
   };
 
@@ -66,23 +69,25 @@ const Button = ({
         { "cursor-not-allowed opacity-80": disabled },
         className
       )}
+      type={type}
     >
       {leftIcon && <span className="mr-2 -ml-1 w-5 h-5">{leftIcon}</span>}
       {icon && <span className="w-5 h-5">{icon}</span>}
-      {children}
+      {loading ? <ClipLoader color="#FF8C4B" /> : children}
+      {loading}
       {rightIcon && <span className="ml-2 -mr-1 w-5 h-5">{rightIcon}</span>}
     </button>
   );
 };
 
-Button.defaultProps = {
-  color: "primary",
-  className: "",
-  leftIcon: null,
-  outline: false,
-  onClick: () => {},
-  rightIcon: null,
-  size: "sm",
-};
+// Button.defaultProps = {
+//   color: "primary",
+//   className: "",
+//   leftIcon: null,
+//   outline: false,
+//   onClick: () => {},
+//   rightIcon: null,
+//   size: "sm",
+// };
 
 export default Button;
