@@ -14,11 +14,11 @@ export function SelectField({
   className,
   containerClassName,
   floatLabel,
-  feedBack,
+  feedBack = "",
   label,
   labelClassName,
   leadingIcon,
-  name,
+  name = "",
   toolTip,
   hintText,
   toolTipTitle,
