@@ -7,7 +7,7 @@ const CallToAction = () => {
     <div className="bg-secondary-25 px-14 py-16 lg:py-20 ">
       <div>
         <div className="mt-2  avatar-group text-center">
-          <div className="relative z-20 inline-flex items-center justify-center w-12 h-12 lg:w-14 lg:h-14 -ml-4  transition-all duration-200 border-2 border-white border-solid rounded-full ease-soft-in-out text-size-xs hover:z-40">
+          <div className="relative z-20 inline-flex items-center justify-center w-12 h-12 lg:w-14 lg:h-14 -ml-4  transition-all duration-200 border-2 border-neutral-white border-solid rounded-full ease-soft-in-out text-size-xs hover:z-40">
             <Image
               src="https://images.unsplash.com/photo-1531427186611-ecfd6d936c79?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8N3x8dXNlcnxlbnwwfHwwfHw%3D&auto=format&fit=crop&w=800&h=800&q=60"
               className="w-full rounded-full"
@@ -16,7 +16,7 @@ const CallToAction = () => {
             />
           </div>
 
-          <div className="relative z-30 inline-flex justify-center w-14 h-14 lg:w-16 lg:h-16 -ml-4  transition-all duration-200 border-2 border-white border-solid rounded-full ease-soft-in-out text-size-xs hover:z-30">
+          <div className="relative z-30 inline-flex justify-center w-14 h-14 lg:w-16 lg:h-16 -ml-4  transition-all duration-200 border-2 border-neutral-white border-solid rounded-full ease-soft-in-out text-size-xs hover:z-30">
             <Image
               src=" https://images.unsplash.com/photo-1573497019940-1c28c88b4f3e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=500&h=500&q=80"
               className="w-full rounded-full"
@@ -25,7 +25,7 @@ const CallToAction = () => {
             />
           </div>
 
-          <div className="relative z-20 inline-flex items-center justify-center w-12 h-12 lg:w-14 lg:h-14 -ml-4  transition-all duration-200 border-2 border-white border-solid rounded-full ease-soft-in-out text-size-xs hover:z-30">
+          <div className="relative z-20 inline-flex items-center justify-center w-12 h-12 lg:w-14 lg:h-14 -ml-4  transition-all duration-200 border-2 border-neutral-white border-solid rounded-full ease-soft-in-out text-size-xs hover:z-30">
             <Image
               src="https://images.unsplash.com/photo-1500648767791-00dcc994a43e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MTd8fHVzZXJ8ZW58MHx8MHx8&auto=format&fit=crop&w=500&h=500&q=60"
               className="w-full rounded-full"
@@ -42,7 +42,7 @@ const CallToAction = () => {
             Can’t find the answers you are looking for?
           </p>
           <Button
-            className="my-3 font-medium text-base md:text-xl font-Inter text-white capitalize py-3.5 px-7"
+            className="my-3 font-medium text-base md:text-xl font-Inter text-neutral-white capitalize py-3.5 px-7"
             size="xl"
           >
             Contact Us Now
