@@ -1,34 +1,23 @@
-import Button from "@/components/form/Button";
 import { WalletAdd, Heart } from "iconsax-react";
+import colors from "@/tokens/color.tokens";
+import { CreateNewTransactType } from "@/components/dashboard/component/card/CreateNewTransact";
 
-const createNew = [
+const createNew: CreateNewTransactType[] = [
   {
-    backGround: "bg-success-25",
-    borderColor: "border-success-100",
+    color: "success",
     Icon: WalletAdd,
-    iconColor: "#00793F",
+    iconColor: colors.success[700],
     title: "Start New Transaction",
-    textColor: "text-success-700",
-    url: "/app/transaction/new",
-    btn: (
-      <Button color="success" size="lg">
-        Lending Transaction
-      </Button>
-    ),
+    url: "/dashboard/transaction/create",
+    btnTitle: "Lending Transaction",
   },
   {
-    backGround: "bg-primary-25",
-    borderColor: "border-primary-100",
+    color: "primary",
     Icon: Heart,
-    iconColor: "#016C95",
+    iconColor: colors.primary[700],
     title: "Quick Start with a Template",
-    textColor: "text-primary-700",
     url: "/app/templates",
-    btn: (
-      <Button color="primary" size="lg">
-        Select a Template
-      </Button>
-    ),
+    btnTitle: "Select a Template",
   },
 ];
 
