@@ -1,3 +1,4 @@
+import { ColorType } from "@/utils/colors";
 import {
   AddCircle,
   UserAdd,
@@ -6,37 +7,35 @@ import {
 } from "iconsax-react";
 
 type newActionProp = {
-  backGround: string;
-  borderColor: string;
+  color: ColorType;
   icon: JSX.Element | React.ReactNode;
   title: string;
   textColor: string;
 };
 const newActiondata: newActionProp[] = [
   {
-    backGround: "bg-primary-25",
-    borderColor: "border-primary-100",
+    color: "primary",
     icon: <AddCircle size="24" variant="Bulk" />,
     title: "Create a New Transaction",
     textColor: "text-primary-600",
   },
   {
-    backGround: "bg-danger-25",
-    borderColor: "border-danger-100",
+    color: "danger",
+
     icon: <UserAdd size="24" variant="Bulk" />,
     title: "Add a New Connection",
     textColor: "text-danger-600",
   },
   {
-    backGround: "bg-success-25",
-    borderColor: "border-success-100",
+    color: "success",
+
     icon: <Profile2User size="24" variant="Bulk" />,
     title: "Create a New Group",
     textColor: "text-success-600",
   },
   {
-    backGround: "bg-warning-25",
-    borderColor: "border-warning-100",
+    color: "warning",
+
     icon: <DocumentUpload size="24" variant="Bulk" />,
     title: "Create a New Template",
     textColor: "text-warning-600",
